user-service/requirements.txt
django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
django-redis>=5.4.0
redis>=5.0.0
gunicorn>=21.2.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
coverage>=7.3.0